Changes in MenuOption.chosenAction() method, files structure refactor, add MenuIndicator with static final values
