+++ src/computer/Main.java
@@ -1,13 +1,15 @@
 package computer;
 
-import computer.hardware.Capacity;
-import computer.hardware.Computer;
-import computer.hardware.Monitor;
-import computer.hardware.drive.Drive;
-import computer.hardware.drive.HDDDrive;
-import computer.hardware.usbdevice.MemoryStick;
-import computer.hardware.usbdevice.Mouse;
-import computer.hardware.usbdevice.USBDevice;
+import computer.hardware.computer.MenuIndicator;
+import computer.software.file.shared.Capacity;
+import computer.hardware.computer.Computer;
+import computer.hardware.components.Monitor;
+import computer.hardware.computer.MenuOption;
+import computer.hardware.components.drive.Drive;
+import computer.hardware.components.drive.HDDDrive;
+import computer.hardware.components.usbdevice.MemoryStick;
+import computer.hardware.components.usbdevice.Mouse;
+import computer.hardware.components.usbdevice.USBDevice;
 import computer.software.file.image.GIFImageFile;
 import computer.software.file.image.JPGImageFIle;
 import computer.software.file.music.MP3File;
@@ -48,7 +50,7 @@ public class Main {
                     3. Hardware
                     4. end <- to exit
                     """);
-            userChoice = MenuOption.chosenAction(scanner.nextLine(),"main");
+            userChoice = MenuOption.chosenAction(scanner.nextLine(), MenuIndicator.MAIN_MENU);
 
             switch (userChoice){
 
@@ -62,7 +64,7 @@ public class Main {
                         back <- to go back
                         end <- to exit
                         """);
-                        userChoice = MenuOption.chosenAction(scanner.nextLine(),"usb");
+                        userChoice = MenuOption.chosenAction(scanner.nextLine(),MenuIndicator.USB_MENU);
 
                         switch (userChoice){
                             case ADD_USB_DEVICE ->{
@@ -101,7 +103,7 @@ public class Main {
                         back <- to go back
                         end <- to exit
                         """);
-                        userChoice = MenuOption.chosenAction(scanner.nextLine(),"file");
+                        userChoice = MenuOption.chosenAction(scanner.nextLine(),MenuIndicator.FILE_MENU);
 
                         switch (userChoice){
                             case ADD_FILE ->{
@@ -142,7 +144,7 @@ public class Main {
                         back <- to go back
                         end <- to exit
                         """);
-                        userChoice = MenuOption.chosenAction(scanner.nextLine(),"hardware");
+                        userChoice = MenuOption.chosenAction(scanner.nextLine(),MenuIndicator.HARDWARE_MENU);
 
                         switch (userChoice){
                             case ADD_HARDWARE ->{
