--- src/computer/hardware/ComponentType.java
@@ -1,6 +1,0 @@
-package computer.hardware;
-
-public enum ComponentType {
-
-    CPU, DRIVE, MONITOR, KEYBOARD, MOUSE, USB, SPEAKER, HEADPHONES, UNKNOWN;
-}
