--- src/computer/hardware/Components.java
@@ -1,7 +1,0 @@
-package computer.hardware;
-
-public interface Components {
-
-    String getName();
-    ComponentType getType();
-}
