+++ src/computer/hardware/computer/MenuIndicator.java
@@ -1,0 +1,9 @@
+package computer.hardware.computer;
+
+public class MenuIndicator {
+    public static final String MAIN_MENU = "Main Menu";
+    public static final String USB_MENU = "USB Menu";
+    public static final String FILE_MENU = "File Menu";
+    public static final String HARDWARE_MENU = "Hardware Menu";
+    public static final String UNIVERSAL = "Universal";
+}
