--- src/computer/hardware/usbdevice/MemoryStickDisconnectedException.java
@@ -1,7 +1,0 @@
-package computer.hardware.usbdevice;
-
-public class MemoryStickDisconnectedException extends Exception {
-    public MemoryStickDisconnectedException(String message) {
-        super(message);
-    }
-}
