--- src/computer/hardware/usbdevice/Mouse.java
@@ -1,12 +1,0 @@
-package computer.hardware.usbdevice;
-
-import computer.hardware.ComponentType;
-
-public class Mouse extends USBDevice{
-
-    // TODO zrobić bliźniaczą klase reprezentującą klawiaturę
-
-    public Mouse(String name) {
-        super(name);
-    }
-}
