--- src/computer/software/file/File.java
@@ -1,7 +1,0 @@
-package computer.software.file;
-
-public interface File {
-    String getName();
-    int getSize();
-    FileType getType();
-}
