--- src/computer/software/file/FileNotFoundException.java
@@ -1,7 +1,0 @@
-package computer.software.file;
-
-public class FileNotFoundException extends Exception {
-    public FileNotFoundException(String message) {
-        super(message);
-    }
-}
