--- src/computer/software/file/FileType.java
@@ -1,5 +1,0 @@
-package computer.software.file;
-
-public enum FileType {
-    MUSIC, IMAGE
-}
