--- src/computer/software/file/NotEnoughSpaceException.java
@@ -1,7 +1,0 @@
-package computer.software.file;
-
-public class NotEnoughSpaceException extends Exception {
-    public NotEnoughSpaceException(String message) {
-        super(message);
-    }
-}
