+++ src/computer/software/file/music/MusicFile.java
@@ -1,6 +1,6 @@
 package computer.software.file.music;
 
-import computer.software.file.File;
+import computer.software.file.shared.File;
 
 public interface MusicFile extends File {
     void play();
